+++ lib/ts/recipe/emailverification/api/implementation.ts
@@ -4,6 +4,7 @@ import EmailVerificationRecipe from "../recipe";
 import { GeneralErrorResponse } from "../../../types";
 import { EmailVerificationClaim } from "../emailVerificationClaim";
 import SessionError from "../../session/error";
+import { getEmailVerifyLink } from "../utils";
 
 export default function getAPIInterface(): APIInterface {
     return {
@@ -102,20 +103,30 @@ export default function getAPIInterface(): APIInterface {
                 });
 
                 if (response.status === "EMAIL_ALREADY_VERIFIED_ERROR") {
+                    if ((await session.getClaimValue(EmailVerificationClaim)) !== true) {
+                        // this can happen if the email was verified in another browser
+                        // and this session is still outdated - and the user has not
+                        // called the get email verification API yet.
+                        await session.fetchAndSetClaim(EmailVerificationClaim, userContext);
+                    }
                     logDebugMessage(
                         `Email verification email not sent to ${emailInfo.email} because it is already verified.`
                     );
                     return response;
                 }
 
-                let emailVerifyLink =
-                    options.appInfo.websiteDomain.getAsStringDangerous() +
-                    options.appInfo.websiteBasePath.getAsStringDangerous() +
-                    "/verify-email" +
-                    "?token=" +
-                    response.token +
-                    "&rid=" +
-                    options.recipeId;
+                if ((await session.getClaimValue(EmailVerificationClaim)) !== false) {
+                    // this can happen if the email was unverified in another browser
+                    // and this session is still outdated - and the user has not
+                    // called the get email verification API yet.
+                    await session.fetchAndSetClaim(EmailVerificationClaim, userContext);
+                }
+
+                let emailVerifyLink = getEmailVerifyLink({
+                    appInfo: options.appInfo,
+                    token: response.token,
+                    recipeId: options.recipeId,
+                });
 
                 logDebugMessage(`Sending email verification email to ${emailInfo}`);
                 await options.emailDelivery.ingredientInterfaceImpl.sendEmail({