+++ lib/ts/recipe/session/sessionClass.ts
@@ -48,7 +48,7 @@ export default class Session implements SessionContainerInterface {
         return this.recipeUserId;
     }
 
-    revokeSession = async (userContext?: any) => {
+    async revokeSession(userContext?: any) {
         await this.helpers.getRecipeImpl().revokeSession({
             sessionHandle: this.sessionHandle,
             userContext: userContext === undefined ? {} : userContext,
@@ -61,9 +61,9 @@ export default class Session implements SessionContainerInterface {
         // returns true, it can cause this kind of a bug:
         // https://github.com/supertokens/supertokens-node/issues/343
         clearSessionFromCookie(this.helpers.config, this.res);
-    };
+    }
 
-    getSessionData = async (userContext?: any): Promise<any> => {
+    async getSessionData(userContext?: any): Promise<any> {
         let sessionInfo = await this.helpers.getRecipeImpl().getSessionInformation({
             sessionHandle: this.sessionHandle,
             userContext: userContext === undefined ? {} : userContext,
@@ -75,9 +75,9 @@ export default class Session implements SessionContainerInterface {
             });
         }
         return sessionInfo.sessionData;
-    };
+    }
 
-    updateSessionData = async (newSessionData: any, userContext?: any) => {
+    async updateSessionData(newSessionData: any, userContext?: any) {
         if (
             !(await this.helpers.getRecipeImpl().updateSessionData({
                 sessionHandle: this.sessionHandle,
@@ -90,25 +90,26 @@ export default class Session implements SessionContainerInterface {
                 type: STError.UNAUTHORISED,
             });
         }
-    };
+    }
 
-    getUserId = (_userContext?: any) => {
+    getUserId(_userContext?: any) {
         return this.userId;
-    };
+    }
 
-    getAccessTokenPayload = (_userContext?: any) => {
+    getAccessTokenPayload(_userContext?: any) {
         return this.userDataInAccessToken;
-    };
+    }
 
-    getHandle = () => {
+    getHandle() {
         return this.sessionHandle;
-    };
+    }
 
-    getAccessToken = () => {
+    getAccessToken() {
         return this.accessToken;
-    };
+    }
 
-    mergeIntoAccessTokenPayload = async (accessTokenPayloadUpdate: any, userContext?: any) => {
+    // Any update to this function should also be reflected in the respective JWT version
+    async mergeIntoAccessTokenPayload(accessTokenPayloadUpdate: any, userContext?: any): Promise<void> {
         const updatedPayload = { ...this.getAccessTokenPayload(userContext), ...accessTokenPayloadUpdate };
         for (const key of Object.keys(accessTokenPayloadUpdate)) {
             if (accessTokenPayloadUpdate[key] === null) {
@@ -117,9 +118,9 @@ export default class Session implements SessionContainerInterface {
         }
 
         await this.updateAccessTokenPayload(updatedPayload, userContext);
-    };
+    }
 
-    getTimeCreated = async (userContext?: any): Promise<number> => {
+    async getTimeCreated(userContext?: any): Promise<number> {
         let sessionInfo = await this.helpers.getRecipeImpl().getSessionInformation({
             sessionHandle: this.sessionHandle,
             userContext: userContext === undefined ? {} : userContext,
@@ -131,9 +132,9 @@ export default class Session implements SessionContainerInterface {
             });
         }
         return sessionInfo.timeCreated;
-    };
+    }
 
-    getExpiry = async (userContext?: any): Promise<number> => {
+    async getExpiry(userContext?: any): Promise<number> {
         let sessionInfo = await this.helpers.getRecipeImpl().getSessionInformation({
             sessionHandle: this.sessionHandle,
             userContext: userContext === undefined ? {} : userContext,
@@ -145,12 +146,14 @@ export default class Session implements SessionContainerInterface {
             });
         }
         return sessionInfo.expiry;
-    };
+    }
 
-    assertClaims = async (claimValidators: SessionClaimValidator[], userContext?: any): Promise<void> => {
+    // Any update to this function should also be reflected in the respective JWT version
+    async assertClaims(claimValidators: SessionClaimValidator[], userContext?: any): Promise<void> {
         let validateClaimResponse = await this.helpers.getRecipeImpl().validateClaims({
             accessTokenPayload: this.getAccessTokenPayload(userContext),
             userId: this.getUserId(userContext),
+            recipeUserId: this.getRecipeUserId(userContext),
             claimValidators,
             userContext,
         });
@@ -166,31 +169,35 @@ export default class Session implements SessionContainerInterface {
                 payload: validateClaimResponse.invalidClaims,
             });
         }
-    };
+    }
 
-    fetchAndSetClaim = async <T>(claim: SessionClaim<T>, userContext?: any) => {
-        const update = await claim.build(this.getUserId(userContext), userContext);
+    // Any update to this function should also be reflected in the respective JWT version
+    async fetchAndSetClaim<T>(claim: SessionClaim<T>, userContext?: any): Promise<void> {
+        const update = await claim.build(this.getUserId(userContext), this.getRecipeUserId(userContext), userContext);
         return this.mergeIntoAccessTokenPayload(update, userContext);
-    };
+    }
 
-    setClaimValue = <T>(claim: SessionClaim<T>, value: T, userContext?: any) => {
+    // Any update to this function should also be reflected in the respective JWT version
+    setClaimValue<T>(claim: SessionClaim<T>, value: T, userContext?: any): Promise<void> {
         const update = claim.addToPayload_internal({}, value, userContext);
         return this.mergeIntoAccessTokenPayload(update, userContext);
-    };
+    }
 
-    getClaimValue = async <T>(claim: SessionClaim<T>, userContext?: any) => {
+    // Any update to this function should also be reflected in the respective JWT version
+    async getClaimValue<T>(claim: SessionClaim<T>, userContext?: any) {
         return claim.getValueFromPayload(await this.getAccessTokenPayload(userContext), userContext);
-    };
+    }
 
-    removeClaim = (claim: SessionClaim<any>, userContext?: any) => {
+    // Any update to this function should also be reflected in the respective JWT version
+    removeClaim(claim: SessionClaim<any>, userContext?: any): Promise<void> {
         const update = claim.removeFromPayloadByMerge_internal({}, userContext);
         return this.mergeIntoAccessTokenPayload(update, userContext);
-    };
+    }
 
     /**
      * @deprecated Use mergeIntoAccessTokenPayload
      */
-    updateAccessTokenPayload = async (newAccessTokenPayload: any, userContext: any) => {
+    async updateAccessTokenPayload(newAccessTokenPayload: any, userContext: any): Promise<void> {
         let response = await this.helpers.getRecipeImpl().regenerateAccessToken({
             accessToken: this.getAccessToken(),
             newAccessTokenPayload,
@@ -218,5 +225,5 @@ export default class Session implements SessionContainerInterface {
                 response.accessToken.expiry
             );
         }
-    };
+    }
 }