+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [12.1.0] - 2022-11-17
+
+### Added:
+
+-   Adds APIs for user details to the dashboard recipe
+
+### Changed:
+
+-   Updates dashboard version to 0.2
+
+## [12.0.6] - 2022-11-09
+
+### Fixes
+
+-   Adds updating of session claims in email verification token generation API in case the session claims are outdated.
+-   Fixed mergeIntoAccessTokenPayload not updating the JWT payload
+
 ## [12.0.5] - 2022-10-17
 
 -   Updated google token endpoint.